requirements.txt
langchain==0.1.12
streamlit==1.33.0
boto3
pdfplumber
python-dotenv
pyinstaller